Adding NUC recovery tests, Wait for SDK - this version is not compilable yet.
